Fix compiler warnings on macOS without changing runtime behavior
- sc1.c: use INT32_MAX/INT32_MIN instead of LONG_MAX/LONG_MIN for 32-bit cell
- sc3.c: suppress unused-but-set warning on debug-only variable nesting
- sc7.c: cast error() return through intptr_t to avoid int-to-pointer warning
- amx_mini.cpp: remove unused variable len
- amxxpc.cpp: use memcpy to avoid taking address of packed struct members

+++ libpc300/sc1.c
@@ -1537,8 +1537,8 @@ static void setconstants(void)
       add_constant("cellmax",_I32_MAX,sGLOBAL,0);
       add_constant("cellmin",_I32_MIN,sGLOBAL,0);
     #else
-      add_constant("cellmax",LONG_MAX,sGLOBAL,0);
-      add_constant("cellmin",LONG_MIN,sGLOBAL,0);
+      add_constant("cellmax",INT32_MAX,sGLOBAL,0);
+      add_constant("cellmin",INT32_MIN,sGLOBAL,0);
     #endif
   #elif PAWN_CELL_SIZE==64
     add_constant("cellbits",64,sGLOBAL,0);
